Allow cache_bench and db_bench to use 3rd party tiered cache
